requirements.txt
flask==3.0.0
openai==1.17.0
azure-ai-ml==1.14.0
python-docx==1.1.0
pytesseract==0.3.10
pillow==10.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
